# openapi/vibe-truth-api-v1.yaml
openapi: 3.1.0
info:
  title: Vibe Truth and Skills API
  version: 1.1.0
  description: Versioned interface to approved Vibe Engineering truth, workflows, context, and deterministic skill execution packets.
servers:
  - url: /api/v1
paths:
  /manifest:
    get:
      operationId: getManifest
      responses:
        '200':
          description: Approved truth manifest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manifest'
  /skills:
    get:
      operationId: listSkills
      responses:
        '200':
          description: Callable Vibe skill catalog
          content:
            application/json:
              schema:
                type: object
                required: [skills]
                properties:
                  skills:
                    type: array
                    items: { $ref: '#/components/schemas/SkillSummary' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
  /skills/{id}:
    get:
      operationId: getSkill
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: One Vibe skill definition
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Skill' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
  /run-skill:
    post:
      operationId: runSkill
      description: Return a deterministic procedure packet. The API does not perform external side effects on behalf of the caller.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunSkillRequest' }
      responses:
        '200':
          description: Skill execution packet
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunSkillResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
  /truth/{id}:
    get:
      operationId: getTruth
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Approved truth artifact }
        '404': { $ref: '#/components/responses/NotFound' }
  /workflows/{id}:
    get:
      operationId: getWorkflow
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Approved workflow artifact }
        '404': { $ref: '#/components/responses/NotFound' }
  /resolve-context:
    post:
      operationId: resolveContext
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextRequest'
      responses:
        '200': { description: Resolved work contract }
        '400': { $ref: '#/components/responses/BadRequest' }
components:
  schemas:
    ArtifactSummary:
      type: object
      additionalProperties: false
      required: [id, kind, version, status, precedence, title, sourcePath, summary, contentHash]
      properties:
        id: { type: string }
        kind: { type: string }
        version: { type: string }
        status: { type: string, enum: [approved] }
        precedence: { type: integer, minimum: 0 }
        title: { type: string }
        sourcePath: { type: string }
        summary: { type: string }
        contentHash: { type: string, pattern: '^[a-f0-9]{64}$' }
    Manifest:
      type: object
      required: [schemaVersion, compilerVersion, sourceCommit, bundleHash, artifacts]
      properties:
        schemaVersion: { type: string }
        compilerVersion: { type: string }
        sourceCommit: { type: string }
        bundleHash: { type: string, pattern: '^[a-f0-9]{64}$' }
        artifacts:
          type: array
          items: { $ref: '#/components/schemas/ArtifactSummary' }
    SkillSummary:
      type: object
      required: [id, title, summary, invocation, category, aliases, outputs, attribution, stepCount]
      properties:
        id: { type: string }
        title: { type: string }
        summary: { type: string }
        invocation: { type: string }
        category: { type: string }
        aliases: { type: array, items: { type: string } }
        outputs: { type: array, items: { type: string } }
        attribution: { type: string }
        stepCount: { type: integer, minimum: 1 }
    Skill:
      allOf:
        - $ref: '#/components/schemas/SkillSummary'
        - type: object
          required: [steps]
          properties:
            steps: { type: array, minItems: 1, items: { type: string } }
    RunSkillRequest:
      type: object
      additionalProperties: false
      required: [id]
      properties:
        id: { type: string, minLength: 1 }
        input: { type: object, additionalProperties: true }
    RunSkillResponse:
      type: object
      required: [skill, input, execution]
      properties:
        skill: { $ref: '#/components/schemas/Skill' }
        input: { type: object, additionalProperties: true }
        execution:
          type: object
          required: [mode, instructions, completionRule]
          properties:
            mode: { type: string, const: agent-procedure }
            instructions: { type: array, minItems: 1, items: { type: string } }
            completionRule: { type: string }
            next: { type: [string, 'null'] }
    ContextRequest:
      type: object
      additionalProperties: false
      properties:
        agent:
          type: object
          additionalProperties: true
        project:
          type: object
          additionalProperties: false
          properties:
            repository: { type: string, pattern: '^[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+$' }
            mode: { type: string, enum: [greenfield, brownfield, discovery, recovery, public-release] }
        task:
          type: object
          additionalProperties: false
          properties:
            type: { type: string, maxLength: 100 }
            requestedAction: { type: string, enum: [inspect, software-change, public-release, destructive-change, content-change, security-review] }
            consequenceLevel: { type: string, enum: [low, medium, high] }
        scope:
          type: object
          additionalProperties: false
          properties:
            allowed: { type: array, items: { type: string } }
            prohibited: { type: array, items: { type: string } }
  responses:
    NotFound:
      description: Resource not found
    BadRequest:
      description: Invalid request
    MethodNotAllowed:
      description: Method not allowed for this resource